requirements.txt
fastapi
uvicorn
motor
pydantic
numpy
joblib
pillow
google-generativeai
python-multipart
scikit-learn
pandas